move inventory and import to own data logic classes
