+++ kriolos-opos-domain/src/main/java/com/openbravo/pos/forms/DataLogicSystem.java
@@ -529,234 +529,7 @@ public class DataLogicSystem extends BeanFactoryDataSingle {
         return (String) m_locationfind.find(iLocation);
     }
 
-//// <editor-fold defaultstate="collapsed" desc="START OF CVIMPORT">     
-
-    /**
-     *
-     * @param csv
-     * @throws BasicException
-     */
-    public final void execCSVStockUpdate(Object[] csv) throws BasicException {
-        //  Push Product Quantity Update into CSVImport table      
-        final SentenceExec m_insertStockUpdateCSVEntry = new StaticSentence(this.session,
-                "INSERT INTO csvimport ( "
-                + "ID, ROWNUMBER, CSVERROR, REFERENCE, CODE, PRICEBUY ) "
-                + "VALUES (?, ?, ?, ?, ?, ?)",
-                new SerializerWriteBasic(new Datas[]{
-            Datas.STRING,
-            Datas.STRING,
-            Datas.STRING,
-            Datas.STRING,
-            Datas.STRING,
-            Datas.DOUBLE
-        }));
-        m_insertStockUpdateCSVEntry.exec(csv);
-
-    }
-
-    /**
-     *
-     * @param csv
-     * @throws BasicException
-     */
-    public final void execAddCSVEntry(Object[] csv) throws BasicException {
-        //  Push Products into CSVImport table 
-        final SentenceExec m_insertCSVEntry = new StaticSentence(this.session,
-                "INSERT INTO csvimport ( "
-                + "ID, ROWNUMBER, CSVERROR, REFERENCE, "
-                + "CODE, NAME, PRICEBUY, PRICESELL, "
-                + "PREVIOUSBUY, PREVIOUSSELL, CATEGORY, TAX) "
-                + "VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)",
-                new SerializerWriteBasic(new Datas[]{
-            Datas.STRING,
-            Datas.STRING,
-            Datas.STRING,
-            Datas.STRING,
-            Datas.STRING,
-            Datas.STRING,
-            Datas.DOUBLE,
-            Datas.DOUBLE,
-            Datas.DOUBLE,
-            Datas.DOUBLE,
-            Datas.STRING,
-            Datas.STRING
-        }));
-        m_insertCSVEntry.exec(csv);
-
-    }
-
-    /**
-     *
-     * @param csv
-     * @throws BasicException
-     */
-    public final void execCustomerAddCSVEntry(Object[] csv) throws BasicException {
-        //  Push Customers into CSVImport table      
-        final SentenceExec m_insertCustomerCSVEntry = new StaticSentence(this.session,
-                "INSERT INTO csvimport ( "
-                + "ID, ROWNUMBER, CSVERROR, SEARCHKEY, NAME) "
-                + "VALUES (?, ?, ?, ?, ?)",
-                new SerializerWriteBasic(new Datas[]{
-            Datas.STRING,
-            Datas.STRING,
-            Datas.STRING,
-            Datas.STRING,
-            Datas.STRING
-        }));
-        m_insertCustomerCSVEntry.exec(csv);
-
-    }
-//// </editor-fold>  
-    
-// This is used by CSVimport to detect what type of product insert we are looking at, or what error occured
-    /**
-     *
-     * @param myProduct Object[]. The array is [0]
-     * @return
-     * @throws BasicException
-     */
-    public final String getProductRecordType(Object[] myProduct) throws BasicException {
-
-        final SentenceFind m_getProductAllFields;
-        final SentenceFind m_getProductRefAndCode;
-        final SentenceFind m_getProductRefAndName;
-        final SentenceFind m_getProductCodeAndName;
-        final SentenceFind m_getProductByReference;
-        final SentenceFind m_getProductByCode;
-        final SentenceFind m_getProductByName;
-
-        m_getProductAllFields = new PreparedSentence(this.session,
-                "SELECT ID FROM products WHERE REFERENCE=? AND CODE=? AND NAME=? ",
-                new SerializerWriteBasic(new Datas[]{Datas.STRING, Datas.STRING, Datas.STRING}),
-                new ProductIdRead()
-        );
-
-        m_getProductRefAndCode = new PreparedSentence(this.session,
-                "SELECT ID FROM products WHERE REFERENCE=? AND CODE=?",
-                new SerializerWriteBasic(new Datas[]{Datas.STRING, Datas.STRING}),
-                new ProductIdRead()
-        );
-
-        m_getProductRefAndName = new PreparedSentence(this.session,
-                "SELECT ID FROM products WHERE REFERENCE=? AND NAME=? ",
-                new SerializerWriteBasic(new Datas[]{Datas.STRING, Datas.STRING}),
-                new ProductIdRead()
-        );
-
-        m_getProductCodeAndName = new PreparedSentence(this.session,
-                "SELECT ID FROM products WHERE CODE=? AND NAME=? ",
-                new SerializerWriteBasic(new Datas[]{Datas.STRING, Datas.STRING}),
-                new ProductIdRead()
-        );
-
-        m_getProductByReference = new PreparedSentence(this.session,
-                "SELECT ID FROM products WHERE REFERENCE=? ",
-                SerializerWriteString.INSTANCE,
-                new ProductIdRead()
-        );
-
-        m_getProductByCode = new PreparedSentence(this.session,
-                "SELECT ID FROM products WHERE CODE=? ",
-                SerializerWriteString.INSTANCE,
-                new ProductIdRead()
-        );
-
-        m_getProductByName = new PreparedSentence(this.session,
-                "SELECT ID FROM products WHERE NAME=? ",
-                SerializerWriteString.INSTANCE,
-                new ProductIdRead()
-        );
-
-        // check if the product exist with all the details, if so return product ID
-        if (m_getProductAllFields.find(myProduct) != null) {
-            return m_getProductAllFields.find(myProduct).toString();
-        }
-        // check if the product exists with matching reference and code, but a different name
-        if (m_getProductRefAndCode.find(myProduct[0], myProduct[1]) != null) {
-            return "Name change";
-        }
-
-        if (m_getProductRefAndName.find(myProduct[0], myProduct[2]) != null) {
-            return "Barcode change";
-        }
-
-        if (m_getProductCodeAndName.find(myProduct[1], myProduct[2]) != null) {
-            return "Reference change";
-        }
-
-        if (m_getProductByReference.find(myProduct[0]) != null) {
-            return "Duplicate Reference found.";
-        }
-
-        if (m_getProductByCode.find(myProduct[1]) != null) {
-            return "Duplicate Barcode found.";
-        }
-
-        if (m_getProductByName.find(myProduct[2]) != null) {
-            return "Duplicate Description found.";
-        }
-
-        return "new";
-    }
-
-    /**
-     *
-     * @param myCustomer
-     * @return
-     * @throws BasicException
-     */
-    public final String getCustomerRecordType(Object[] myCustomer) throws BasicException {
-
-        final SerializerRead customerIdRead = (DataRead dr) -> (dr.getString(1));
-
-        final SentenceFind m_getCustomerAllFields;
-        final SentenceFind m_getCustomerSearchKeyAndName;
-        final SentenceFind m_getCustomerBySearchKey;
-        final SentenceFind m_getCustomerByName;
-
-        // duplicate this for now as will extend in future release 
-        m_getCustomerAllFields = new PreparedSentence(this.session,
-                "SELECT ID FROM customers WHERE SEARCHKEY=? AND NAME=? ",
-                new SerializerWriteBasic(new Datas[]{Datas.STRING, Datas.STRING}),
-                customerIdRead
-        );
-
-        m_getCustomerSearchKeyAndName = new PreparedSentence(this.session,
-                "SELECT ID FROM customers WHERE SEARCHKEY=? AND NAME=? ",
-                new SerializerWriteBasic(new Datas[]{Datas.STRING, Datas.STRING}),
-                customerIdRead
-        );
-
-        m_getCustomerBySearchKey = new PreparedSentence(this.session,
-                "SELECT ID FROM customers WHERE SEARCHKEY=? ",
-                SerializerWriteString.INSTANCE,
-                customerIdRead
-        );
-
-        m_getCustomerByName = new PreparedSentence(this.session,
-                "SELECT ID FROM customers WHERE NAME=? ",
-                SerializerWriteString.INSTANCE,
-                customerIdRead
-        );
-
-        if (m_getCustomerAllFields.find(myCustomer) != null) {
-            return m_getCustomerAllFields.find(myCustomer).toString();
-        }
-
-        if (m_getCustomerSearchKeyAndName.find(myCustomer[0], myCustomer[1]) != null) {
-            return "reference error";
-        }
-
-        if (m_getCustomerBySearchKey.find(myCustomer[0]) != null) {
-            return "Duplicate Search Key found.";
-        }
-
-        if (m_getCustomerByName.find(myCustomer[1]) != null) {
-            return "Duplicate Name found.";
-        }
-
-        return "new";
-    }
+//// </editor-fold> 
 
     public final void updatePlaces(int x, int y, String id) throws BasicException {
         final SentenceExec m_updatePlaces = new StaticSentence(this.session,
